# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API 문서",
    "version": "1.0.0",
    "description": ""
  },
  "host": "",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/coin/transfer": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "sop transfer body",
            "schema": {
              "$ref": "#/definitions/TransferReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all success response",
            "schema": {
              "$ref": "#/definitions/TxSuccessRes"
            }
          },
          "500": {
            "description": "all errors",
            "schema": {
              "$ref": "#/definitions/ErrRes"
            }
          }
        }
      }
    },
    "/sop/transfer": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "sop transfer body",
            "schema": {
              "$ref": "#/definitions/TransferReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all success response",
            "schema": {
              "$ref": "#/definitions/TxSuccessRes"
            }
          },
          "500": {
            "description": "all errors",
            "schema": {
              "$ref": "#/definitions/ErrRes"
            }
          }
        }
      }
    },
    "/flux/delegate": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "flux delegate body",
            "schema": {
              "$ref": "#/definitions/FluxDelegateReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all success response",
            "schema": {
              "$ref": "#/definitions/TxSuccessRes"
            }
          },
          "500": {
            "description": "all errors",
            "schema": {
              "$ref": "#/definitions/ErrRes"
            }
          }
        }
      }
    },
    "/flux/undelegate": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "flux undelegate body",
            "schema": {
              "$ref": "#/definitions/FluxUndelegateReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all success response",
            "schema": {
              "$ref": "#/definitions/TxSuccessRes"
            }
          },
          "500": {
            "description": "all errors",
            "schema": {
              "$ref": "#/definitions/ErrRes"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "TransferReq": {
      "type": "object",
      "properties": {
        "keyID": {
          "type": "string",
          "example": "2f4c980d-5e3b-4139-8e85-a3c154190876",
          "describe": "kms keyID"
        },
        "userAddr": {
          "type": "string",
          "example": "0x053495aB8dD181c6413be812fE63F56Af43a625E",
          "describe": "to address"
        },
        "to": {
          "type": "string",
          "example": "0x5F90d10443B03F46a6c3513fe62F60733E7BceA7",
          "describe": "to address"
        },
        "amount": {
          "type": "string",
          "example": 1,
          "describe": "eth unit amount"
        }
      }
    },
    "FluxDelegateReq": {
      "type": "object",
      "properties": {
        "keyID": {
          "type": "string",
          "example": "2f4c980d-5e3b-4139-8e85-a3c154190876",
          "describe": "kms keyID"
        },
        "userAddr": {
          "type": "string",
          "example": "0x053495aB8dD181c6413be812fE63F56Af43a625E",
          "describe": "to address"
        },
        "soID": {
          "type": "string",
          "example": "12",
          "describe": "so ID"
        },
        "amount": {
          "type": "string",
          "example": 1,
          "describe": "eth unit amount"
        }
      }
    },
    "FluxUndelegateReq": {
      "type": "object",
      "properties": {
        "keyID": {
          "type": "string",
          "example": "2f4c980d-5e3b-4139-8e85-a3c154190876",
          "describe": "kms keyID"
        },
        "userAddr": {
          "type": "string",
          "example": "0x053495aB8dD181c6413be812fE63F56Af43a625E",
          "describe": "to address"
        },
        "soID": {
          "type": "string",
          "example": "12",
          "describe": "so ID"
        }
      }
    },
    "ErrRes": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "failed"
        }
      }
    },
    "TxSuccessRes": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "success"
        },
        "txHash": {
          "type": "string",
          "example": "0xdf62d376699d78c736b55186dfe27bbb2f60b423ba00b753b690cfb4d7d62008"
        }
      }
    }
  }
}